Updated canvas drawing functions
This comment can be found in view-source.ts:211 and is on about the WebGL rendering method I was testing:
UPDATE/CONCLUSION: So I got this working, there were a few visual bugs that I didn't like, I will probably come back to that.
But anyway, which runs better, the WebGL context or 2D context? From my testing and with my code, 2D, the WebGL one RUNS LIKE ASS in comparison.
I wont be using WebGL for now until I optimize my code.
It's a little annoying that it didn't work as well as I had hoped considering the amount of time I spend looking into WebGL, but at least I learnt some things about it and it could be handy in the future.
I may come back to this WebGL stuff and try and make it better but for now I will be using the 2D context.
I will be leaving this WebGL stuff in here but commented out, this project is very experimental anyway. The C# stuff is also quite messy, the NDI stuff I made also isn't used.